requirements.txt
jupyter~=1.1.1
ipython~=8.31.0
requests~=2.32.3
aiogram~=3.17.0
python-dotenv~=1.0.1
loguru~=0.7.3
